rename species NA -> NL
In this commit, I moved the original dataset into a subfolder (`data_orig/`). I
wrote a simple R script that renames the species from NA to NL (and some limited
tests comparing the original and the new datasets). The modified dataset is
written where it was originally.

There are few other changes to the files that shouldn't impact anything else:
- the headers are now quoted
- columns that only contain numbers are not quoted

I have no problem to have this script (or other variation posted on figshare or
wherever Morgan would prefer).