Update 03-dates-as-data.md
"we can easily add days, months or years to a given date." https://github.com/fishtree-attempt/spreadsheet-ecology-lesson/blob/main/episodes/03-dates-as-data.md?plain=1#L133

then just a few lines below it reads: 

"Adding years and months and days is slightly trickier because we need to make
sure that we are adding the amount to the correct entity."


- First we extract the single entities (day, month or year)
- We can then add values to do that
- Finally the complete date string is reconstructed using the `DATE()` function."
https://github.com/fishtree-attempt/spreadsheet-ecology-lesson/blob/main/episodes/03-dates-as-data.md?plain=1#L154-L159


So: Firstly it is stated that adding days/months/years to a given date is easy while the same task ("Adding years and months and days") is deemed "slightly trickier".

The first bullet asks to extract the single entities. I am not sure what this whole note is about. I feel that adding years to another date might be consdiered "tricky"? Why is that?

The second bullet is very unspecific. What is "that"?

The third bullet talks about a reconstruction.

What is the main challenge?
Was the date not formatted correctly in the first instance and that's why adding days, etc. is complicated?
This whole 'Note' confuses me.
We should add an example that clearly distinguishes the easy approach (#L133) from the (#L154) 'trickier' version. 

If others do not see the point (I don't), we could also consider deleting this whole note?